+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Bring back handleSubmit when using submit button. Fixes UIU-743.
 * Fix permission menu after reopening. Fixes UIU-739.
 * Apply internationalization to all hardcoded strings for Settings. Fixes UIU-691.
+* Pass `userBarcode` to new requests. Fixes UIU-690. Refs UIREQ-146.
 
 ## [2.17.0](https://github.com/folio-org/ui-users/tree/v2.17.0) (2018-10-5)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.16.0...v2.17.0)